:muscle: Add `Plugin` layer to simplify code
