+++ denops/@denops-private/plugin/base.ts
@@ -1,0 +1,6 @@
+export type Plugin = {
+  readonly name: string;
+  readonly script: string;
+  call(fn: string, ...args: unknown[]): Promise<unknown>;
+  dispose(): void;
+};
