+++ denops/@denops-private/plugin/worker/patch_console.ts
